pyproject.toml
[project]
name = "breba-proxy"
version = "0.1.0"
description = "FastAPI app with wildcard subdomain support"

authors = [{ name = "Yason Khaburzaniya", email = "yason@hey.com" }]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv>=1.1.0",
    "boto3>=1.40.27",
]

[tool.uv]

